Split gdb_stub.c into organized files in its own directory
