+++ kernel/arch/dreamcast/kernel/gdb/Makefile
@@ -1,0 +1,9 @@
+# KallistiOS ##version##
+#
+# arch/dreamcast/kernel/gdb/Makefile
+# Copyright (C) 2026 Andy Barajas
+#
+
+OBJS = gdb.o gdb_ctrl.o gdb_regs.o gdb_mem.o gdb_break.o gdb_packet.o gdb_utils.o
+
+include $(KOS_BASE)/Makefile.prefab
